Refactor GetManager to deduce the document type
The document type is embedded in the documentID so we should test the
documentID directly for the document type. I had to create a special
exception for the "SCTEST" document type so we could use test documents
in our integration tests. The "SCTEST" document type is only considered
valid if a specific environment variable is present.