Properly clean up manager instances
The API service needs to check for nil errors so we don't try to
dereference the nil pointer. We also need to call mgr.Close() in all
cases before we continue the loop. It would be interesting to look into
using an anonymous function for this. However, manually handling
manager cleanup is a bit easier for now.