Remove the job to cleanup untagged images
Docker updated the way images are pushed into registries and the
snok/container-retention-policy action hasn't been updated to account
for that change. This broke our container registry in GHCR for images
that we want to retain. The easiest way to fix this is to remove the
job that cleans up untagged images for now.